Setup fixture directory for tests, and make it discoverable
